Question 3
For question 3, I created a bullet prefab with a script called BulletScript. This script is responsible for adding an impulse force to the bullet's rigidbody to give the bullet-like velocity. The script has a function that is called by the player controller after instantiation of the bullet prefab that determines direction based on the player's last input (determines where the player is facing). The bullet's damage is also determined by the player script and can be changed in the Unity editor.  When an enemy collides with any object with the bullet tag, the enemy's health decreases based on its given damage level. I changed the health script as well to take an optional health percentage decrease variable in order to be able to change the amount of damage taken by an interaction/collision, and the bullet's damage variable is inputted for that value when colliding with an enemy.