Question 1
In order to give variation in health that can be decreased based on individual enemies, I gave each enemy a public damage value that is called by the PlayerEnemyCollision event script. I also used the earlier change that I made to the health system that takes an optional health damage multiplier. By changing the code that runs upon PlayerEnemyCollision that is not within the bounds of destroying the enemy, I was able to implement the Decrement function on the player's health value based on the Enemy's damage power. Now it takes multiple collisions with the enemy to kill the player. Due to time constraints, I'll later attempt to create some type of UI feedback system.