Question 2
I implemented a timer within the PlayerController that switches a bool called immunityOn to true for 1 second. While immunityOn is true, the PlayerEnemyCollision script prevents collisions between the Player and Enemy to cause loss of health in the Player. The timer is reset on every collision between the two with a function called StartImmunity.