2 Hour Mark Update
I logged about two hours for Question 1-3. I wasn't able to add a visual UI feedback for Question 1, so that one is technically a WIP. Question 2 & 3 are complete. I wanted to get some clarification on Question 4 before proceeding, and though I am over the two hour limit, I'd like to still try the 4th question if possible. I'll log the time worked on that question as well. The explanations for each question are in the description of each commit.