Question 4
For question 4, I felt the simplest solution to make the gamemode creation friendly and simple for the game designer was to create a prefab that could be dragged and dropped into the scene and have its values edited. I created the script for the GamemodeTemplate with the same name that takes a transform spawnpoint and a float MaxHP. I then went in and made the "Model" script take a GamemodeTemplate as a variable, making the various Gamemode's and their parameters acccessible by the PlayerSpawn script. This way when a player spawns into the scene or upon respawning, the Player will have health and a spawnpoint according to the gamemode. All a designer must do to create various gamemodes is to drag and drop a GamemodeTemplate prefab into the scene, edit the values on the GamemodeTemplate script, and then drag it into place in the Model Script under the GameController object.