Question 1 / Question 2
In the Health script, I replaced the decrement method with a TakeDamage method for simplicity and neater code. I also added a method that updates the HealthUI. I then changed the PlayerEnemyCollision script to include the TakeDamage method and set the damage parameter from the takedamage method as 20.  The maxHP is now 100. For question two I implemented a coroutine for immunity in the playerController class rather than the PlayerEnemyCollison class since it affects the player's behaviors rather than another collision.