Question 4
This is either finished or pt1 of what im adding for question 4. For it i added and edited  a scriptable object that changes player health, ability to take damage, invulnerability time, movement speed and jump speed projectile damage, and despawn time, and starting postion.
Future addition if mores requested will be enemy setting and jump damage , more in depth would include ability to fire being on or off aswell.