question 4
Built-in a small UI and created a game mode script and asset that allows the game designer to change the game mode name, current game mode hp, starting position, player speed, bullet damage, and the bullet prefab for some more customization. The main 4 changes modify the way gameplay works for the player and the final change allows for a more fun experience.